requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.2.0
boto3==1.35.10
botocore==1.35.10
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==4.2.15
django-celery-beat==2.7.0
django-environ==0.11.2
django-timezone-field==7.0
djangorestframework==3.15.2
idna==3.8
jmespath==1.0.1
kombu==5.4.0
numpy==1.24.4
opencv-python==4.10.0.84
pandas==2.0.3
pillow==10.4.0
prompt-toolkit==3.0.47
psycopg2-binary==2.9.9
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.8
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
